Json parse refactor (#11)
The existing JSON deserialization created either integers or doubles
depending on whether the number had decimals. In the default JavaScript
serializer I have no control over this, so this caused errors if an
integer was deduced while a double was expected.

The implemented solution is to not do the automatic translation from
JSON to Java built-in objects any more. Instead it constructs the
domain objects from Jacksons JsonNode objects.

Changing the deserialization of policies and template assets was
skipped to save on time.