No square (#21)
Previously the selected buurt or area was reduced to a square bounding
box. With this change the geometry of the bounds are respected.

Internally it still keeps all the panden in the square box but it
filters them out on-demand.

This change took a lot longer than I expected due to some research
and various issues.

Panden and verblijfsobjecten are now stored in a Map instead of an
Array, for de-duplication. I've also switch to using iterators instead
of Arrays for derived data. I think the change to Map was a good thing
but the change to iterators mostly introduced complexity and slowness
for no gain. I might revert it later.