Edit building selection (#178)
Until now you could only select buildings when uploading an excel file.

This change puts the panden select map in the survey form. This enables
selecting buildings in more situations.

I moved some fetch-related code from components to loaders because it
needed to be changed and this is a simpler approach.